(refactor) Use `Fact` and `Intent` message abstractions (#3)
We distinguish between messages that are intended to change the state of the
application which can be verified and rejected if need be and those that state
irrefutable facts. We protect our invariants by only allowing `Intent`s into
the system, stating facts is the sole prerogative of the system itself.

See this discourse thread for more context:
https://discourse.elm-lang.org/t/intents-and-facts-pondering-cqrs-in-elm/923